+++ sbuildr/cli/cli.py
@@ -1,10 +1,9 @@
-from sbuildr.generator.rbuild import RBuildGenerator
 from sbuildr.project.target import ProjectTarget
 from sbuildr.project.project import Project
 from sbuildr.logger import G_LOGGER, plural
+from sbuildr.misc import paths, utils
 from sbuildr.graph.node import Node
 import sbuildr.logger as logger
-from sbuildr.misc import paths
 
 from collections import defaultdict
 from typing import List, Tuple
@@ -28,41 +27,27 @@ elif args.verbose:
 
 # Sets up the the command-line interface for the given project/generator combination.
 # When no profile(s) are specified, default_profile will be used.
-def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=["debug", "release"]):
+def cli(project: Project, default_profiles=["debug", "release"]):
     """
     Adds the SBuildr command-line interface to the Python script invoking this function. For detailed usage information, you can run the Python code invoking this function with ``--help``.
 
     :param project: The project that the CLI will interface with.
-    :param GeneratorType: The type of generator to use for generating configuration files. Since SBuildr is a meta-build system, it can support multiple backends to perform builds. For example, RBuild (i.e. ``sbuildr.generator.RBuildGenerator``) can be used for fast incremental builds.
     :param default_profiles: Names of default profiles. These are the profiles the `build`, `run`, and `clean` targets will use when none are explicitly specified via the command-line. Note that `tests` will run tests for all profiles by default.
     """
-    generator = GeneratorType(project)
-
     def needs_configure(func, *args, **kwargs):
         def wrapper(*args, **kwargs):
-            if generator.needs_configure():
+            if project.needs_configure():
                 G_LOGGER.warning(f"'{func.__name__}' requires the project to be configured, but project has either not been configured yet, or configuration is outdated. Running configuration now.")
                 configure(*args, **kwargs)
             return func(*args, **kwargs)
         return wrapper
 
-    def _wrap_str(inp: str, wrap: str='='):
-        terminal_width, _ = shutil.get_terminal_size()
-        return inp.center(terminal_width, wrap)
 
-    # Returns the captured output
-    # TODO: Subprocess management needs to be centralized somewhere.
-    def _check_returncode(result: subprocess.CompletedProcess) -> str:
-        output = f"{_wrap_str(' Captured stdout ')}\n{result.stdout.decode(sys.stdout.encoding)}\n{_wrap_str(' Captured stderr ')}\n{result.stderr.decode(sys.stdout.encoding)}"
-        if result.returncode:
-            G_LOGGER.critical(f"Failed with:\n{output}\nReconfiguring the project or running a clean build may resolve this.")
-        return output
-
-    def _all_targets() -> List[ProjectTarget]:
+    def all_targets() -> List[ProjectTarget]:
         return list(project.libraries.values()) + list(project.executables.values())
 
     # Given target names, returns the corresponding targets.
-    def _select_targets(args) -> List[ProjectTarget]:
+    def select_targets(args) -> List[ProjectTarget]:
         targets = []
         for tgt_name in args.targets:
             if tgt_name not in project:
@@ -77,86 +62,65 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=
                 G_LOGGER.warning(f"Target: {tgt_name} refers to both an executable and a library. Selecting both.")
         return targets
 
-    def _select_test_targets(args) -> List[ProjectTarget]:
-        targets = []
-        for test_name in args.targets:
-            if test_name not in project.tests:
-                G_LOGGER.critical(f"Could not find test: {test_name} in project.\n\tAvailable tests:\n\t\t{list(project.tests.keys())}")
-            targets.append(project.tests[test_name])
-        return targets or list(project.tests.values())
-
     # Given argparse's args struct, parses out profile flags, and returns a list of profile names included.
-    # Falls back to returning the default profile.
-    def _select_profile_names(args) -> List[str]:
+    def select_profile_names(args) -> List[str]:
         return [prof_name for prof_name in project.profiles.keys() if getattr(args, prof_name)]
 
-    def _select_nodes(targets: List[ProjectTarget], prof_names: List[str]) -> List[Node]:
-        nodes = []
-        # Create all required profile build directories and populate nodes.
-        for prof_name in prof_names:
-            if prof_name not in project.profiles:
-                G_LOGGER.critical(f"Profile {prof_name} does not exist in the project. Available profiles: {list(project.profiles.keys())}")
-            project.files.mkdir(project.profiles[prof_name].build_dir)
-            # Populate nodes.
-            for target in targets:
-                if prof_name in target:
-                    node = target[prof_name]
-                    G_LOGGER.verbose(f"For target: {target}, profile: {prof_name}, found path: {node.path}")
-                    nodes.append(node)
-                else:
-                    G_LOGGER.debug(f"Skipping target: {target.name} for profile: {prof_name}, as it does not exist.")
-        return nodes
 
-    def _build_nodes(nodes: List[Node]) -> float:
-        status, time_elapsed = generator.build(nodes)
-        _check_returncode(status)
-        return time_elapsed
-
-    def _build_targets(targets: List[ProjectTarget], prof_names: List[str]):
-        G_LOGGER.info(f"Building targets: {[target.name for target in targets]} for profiles: {prof_names}")
-        G_LOGGER.debug(f"Targets: {targets}")
-
-        time_elapsed = _build_nodes(_select_nodes(targets, prof_names))
-        G_LOGGER.info(f"Built {plural('target', len(targets))} for {plural('profile', len(prof_names))} in {time_elapsed} seconds.")
-
-    def help_targets(args):
-        targets = _select_targets(args) or _all_targets()
-        G_LOGGER.info(f"\n{_wrap_str(' Targets ')}")
+    def help(args):
+        targets = select_targets(args) or all_targets()
+        G_LOGGER.info(f"\n{utils.wrap_str(' Targets ')}")
         for target in targets:
             G_LOGGER.info(f"Target: {target} {'(internal)' if target.internal else ''}. Available Profiles:")
             for prof, node in target.items():
                 G_LOGGER.info(f"\tProfile: {prof}. Path: {node.path}.")
-        G_LOGGER.info(f"\n{_wrap_str(' Public Interface ')}")
+        G_LOGGER.info(f"\n{utils.wrap_str(' Public Interface ')}")
         G_LOGGER.info(f"Headers: {project.public_headers}")
 
+
     def configure(args):
-        generator.generate()
+        project.configure()
+
 
     @needs_configure
     def build(args) -> Tuple[List[ProjectTarget], List[str]]:
-        targets = _select_targets(args) or _all_targets()
-        prof_names = _select_profile_names(args) or default_profiles
-        _build_targets(targets, prof_names)
+        targets = select_targets(args) or all_targets()
+        prof_names = select_profile_names(args) or default_profiles
+        project.build(targets, prof_names)
+
 
     @needs_configure
     def run(args):
         if args.target not in project.executables:
             G_LOGGER.critical(f"Could not find target: {args.target} in project executables. Note: Available executables are: {list(project.executables.keys())}")
         target = project.executables[args.target]
-        prof_name = _select_profile_names(args)[0] or default_profiles
-        _build_targets([target], [prof_name])
+        prof_name = select_profile_names(args)[0] or default_profiles
+        project.build([target], [prof_name])
         G_LOGGER.log(f"\nRunning target: {target}, for profile: {prof_name}: {target[prof_name].path}", color=logger.Color.GREEN)
-        G_LOGGER.log(_check_returncode(subprocess.run([target[prof_name].path], capture_output=True)))
+        status = subprocess.run([target[prof_name].path], capture_output=True)
+        output = f"{utils.wrap_str(' Captured stdout ')}\n{result.stdout.decode(sys.stdout.encoding)}\n{utils.wrap_str(' Captured stderr ')}\n{result.stderr.decode(sys.stdout.encoding)}"
+        G_LOGGER.log(output)
+        if result.returncode:
+            G_LOGGER.critical(f"Failed to run. Reconfiguring the project or running a clean build may resolve this.")
+
 
     @needs_configure
     def tests(args):
-        tests = _select_test_targets(args)
-        prof_names = _select_profile_names(args) or list(project.profiles.keys())
+        def select_test_targets(args) -> List[ProjectTarget]:
+            targets = []
+            for test_name in args.targets:
+                if test_name not in project.tests:
+                    G_LOGGER.critical(f"Could not find test: {test_name} in project.\n\tAvailable tests:\n\t\t{list(project.tests.keys())}")
+                targets.append(project.tests[test_name])
+            return targets or list(project.tests.values())
+
+        tests = select_test_targets(args)
+        prof_names = select_profile_names(args) or list(project.profiles.keys())
         if not tests:
             G_LOGGER.warning(f"No tests found. Have you registered tests using project.test()?")
             return
         # Otherwise, build and run the specified tests
-        _build_targets(tests, prof_names)
+        project.build(tests, prof_names)
 
         class TestResult:
             def __init__(self):
@@ -166,7 +130,7 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=
         test_results = defaultdict(TestResult)
         failed_targets = defaultdict(set)
         for prof_name in prof_names:
-            G_LOGGER.log(f"\n{_wrap_str(f' Profile: {prof_name} ')}", color=logger.Color.GREEN)
+            G_LOGGER.log(f"\n{utils.wrap_str(f' Profile: {prof_name} ')}", color=logger.Color.GREEN)
             for test_target in tests:
                 G_LOGGER.log(f"\nRunning test: {test_target}, for profile: {prof_name}: {test_target[prof_name].path}\n", color=logger.Color.GREEN)
                 status = subprocess.run([test_target[prof_name].path])
@@ -178,7 +142,7 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=
                     G_LOGGER.log(f"\nPASSED {test_target}", color=logger.Color.GREEN)
                     test_results[prof_name].passed += 1
         # Display summary
-        G_LOGGER.log(f"\n{_wrap_str(f' Test Results Summary ')}\n", color=logger.Color.GREEN)
+        G_LOGGER.log(f"\n{utils.wrap_str(f' Test Results Summary ')}\n", color=logger.Color.GREEN)
         for prof_name, result in test_results.items():
             if result.passed or result.failed:
                 G_LOGGER.log(f"Profile: {prof_name}", color=logger.Color.GREEN)
@@ -187,82 +151,59 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=
                 if result.failed:
                     G_LOGGER.log(f"\tFAILED {plural('test', result.failed)}: {failed_targets[prof_name]}", color=logger.Color.RED)
 
-    # TODO: Need a wrapper that creates symlinks for version.
-    # Copies src to dst. dst may be either a complete path or containing directory.
-    def _copy_file(src, dst) -> bool:
-        try:
-            os.makedirs(os.path.dirname(dst), exist_ok=True)
-            shutil.copy2(src, dst)
-            return True
-        except PermissionError:
-            G_LOGGER.error(f"Could not write to {dst}. Do you have sufficient privileges?")
-            return False
 
-    def _get_install_nodes(args):
-        headers, targets = [], []
-        for tgt in args.targets:
-            if tgt in project:
-                targets.append(tgt)
-            else:
-                headers.append(tgt)
-        args.targets = targets
-        headers = headers or list(project.public_headers)
-        targets = [tgt for tgt in (_select_targets(args) or _all_targets()) if not tgt.internal]
+    def get_install_targets(args):
+        headers = [tgt for tgt in args.targets if tgt not in project] or list(project.public_headers)
+        args.targets = [tgt for tgt in args.targets if tgt in project]
+        targets = [tgt for tgt in (select_targets(args) or all_targets()) if not tgt.internal]
         G_LOGGER.verbose(f"Selected public targets: {targets}")
-        prof_names = _select_profile_names(args) or ["release"]
-        return targets, headers, prof_names
-
-    def _header_install_path(args, header: str):
-        return os.path.join(args.headers, os.path.basename(header))
-
-    # TODO: Add -f flag and --upgrade behavior should be to remove older versions.
-    @needs_configure
-    def install(args):
-        targets, headers, prof_names = _get_install_nodes(args)
-        _build_targets(targets, prof_names)
+        prof_names = select_profile_names(args) or ["release"]
         G_LOGGER.verbose(f"Installing targets: {targets} for profiles: {prof_names}")
         G_LOGGER.verbose(f"Installing headers: {headers}")
+        return targets, prof_names, headers
 
+    def get_install_nodes(args, targets, prof_names, headers):
+        # Maps target nodes to installation paths
+        node_installs: List[Tuple[Node, str]] = []
         for prof_name in prof_names:
             for target in targets:
-                install_dir = args.libraries if target.is_lib else args.executables
                 node = target[prof_name]
+                install_dir = args.libraries if target.is_lib else args.executables
                 install_path = os.path.join(install_dir, node.name)
-                if _copy_file(node.path, install_path):
-                    G_LOGGER.info(f"Installed target: {node.name} to {install_path}")
+                node_installs.append((node, install_path))
 
+        # Maps header paths to installation paths.
+        header_installs: List[Tuple[str, str]] = []
         for header in headers:
             candidates = project.files.find(header)
             if len(candidates) == 0:
                 G_LOGGER.critical(f"Could not find installation header: {header}")
             if len(candidates) > 1:
                 G_LOGGER.critical(f"For installation header: {header}, found multiple installation candidates: {candidates}. Please provide a longer path to disambiguate.")
-            header = candidates[0]
-            install_path = _header_install_path(args, header)
-            if _copy_file(header, install_path):
+            header_installs.append((candidates[0], os.path.join(args.headers, os.path.basename(header))))
+
+        return node_installs, header_installs
+
+
+    # TODO: Add -f flag and --upgrade behavior should be to remove older versions.
+    @needs_configure
+    def install(args):
+        targets, prof_names, headers = get_install_targets(args)
+        project.build(targets, prof_names)
+        node_installs, header_installs = get_install_nodes(args, targets, prof_names, headers)
+        for node, install_path in node_installs:
+            if utils.copy_file(node.path, install_path):
+                G_LOGGER.info(f"Installed target: {node.name} to {install_path}")
+
+        for header, install_path in header_installs:
+            if utils.copy_file(header, install_path):
                 G_LOGGER.info(f"Installed header: {header} to {install_path}")
 
 
     def uninstall(args):
-        targets, headers, prof_names = _get_install_nodes(args)
-        G_LOGGER.verbose(f"Uninstalling targets: {targets} for profiles: {prof_names}")
-        G_LOGGER.verbose(f"Uninstalling headers: {headers}")
-
-        install_paths = []
-        for prof_name in prof_names:
-            for target in targets:
-                install_dir = args.libraries if target.is_lib else args.executables
-                node = target[prof_name]
-                install_paths.append(os.path.join(install_dir, node.name))
-
-        for header in headers:
-            candidates = project.files.find(header)
-            if len(candidates) == 0:
-                G_LOGGER.critical(f"Could not find installation header: {header}")
-            if len(candidates) > 1:
-                G_LOGGER.critical(f"For uninstallation header: {header}, found multiple candidates: {candidates}. Please provide a longer path to disambiguate.")
-            header = candidates[0]
-            install_paths.append(_header_install_path(args, header))
+        targets, prof_names, headers = get_install_targets(args)
+        node_installs, header_installs = get_install_nodes(args, targets, prof_names, headers)
+        install_paths = [path for (_, path) in node_installs] + [path for (_, path) in header_installs]
 
         if not args.force:
             G_LOGGER.warning(f"Uninstall dry-run, will not remove files without -f/--force.")
@@ -277,6 +218,7 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=
                 G_LOGGER.info(f"Would remove: {install_path}")
 
 
+    # TODO: Add -f/--force option without which it will not clean.
     def clean(args):
         # TODO(3): Finish implementation, add per-target cleaning.
         to_remove = []
@@ -286,12 +228,13 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=
             G_LOGGER.info(f"Initiating Nuclear Protocol!")
         else:
             # By default, cleans all targets for the default profile.
-            prof_names = _select_profile_names(args) or default_profiles
+            prof_names = select_profile_names(args) or default_profiles
             to_remove = [project.profiles[prof_name].build_dir for prof_name in prof_names]
             G_LOGGER.info(f"Cleaning targets for profiles: {prof_names}")
         # Remove
         for path in to_remove:
             project.files.rm(path)
+
 
     parser = argparse.ArgumentParser(description="Builds this project", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument("-v", "--verbose", help="Enable verbose logging output", action="store_true")
@@ -305,9 +248,9 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=
     subparsers = parser.add_subparsers()
 
     # List
-    help_targets_parser = subparsers.add_parser("targets", help="Display information about targets", description="Display information about the targets in this project")
-    help_targets_parser.add_argument("targets", nargs='*', help="Targets to display. By default, displays help information for all targets.", default=[])
-    help_targets_parser.set_defaults(func=help_targets)
+    help_parser = subparsers.add_parser("help", help="Display information about available targets and public headers", description="Display information about the targets and public headers in this project")
+    help_parser.add_argument("targets", nargs='*', help="Targets to display. By default, displays help information for all targets.", default=[])
+    help_parser.set_defaults(func=help)
 
     # Configure
     configure_parser = subparsers.add_parser("configure", help="Generate build configuration files", description="Generate a build configuration file for the project")